+++ ansible/lib/ansible/modules/extras/system/osx_defaults.py
@@ -33,6 +33,13 @@ options:
       - The domain is a domain name of the form com.companyname.appname.
     required: false
     default: NSGlobalDomain
+  host:
+    description:
+      - The host on which the preference should apply. The special value "currentHost" corresponds to the
+        "-currentHost" switch of the defaults commandline tool.
+    required: false
+    default: null
+    version_added: "2.1"
   key:
     description:
       - The key of the user preference
@@ -67,6 +74,7 @@ notes:
 EXAMPLES = '''
 - osx_defaults: domain=com.apple.Safari key=IncludeInternalDebugMenu type=bool value=true state=present
 - osx_defaults: domain=NSGlobalDomain key=AppleMeasurementUnits type=string value=Centimeters state=present
+- osx_defaults: domain=com.apple.screensaver host=currentHost key=showClock type=int value=1
 - osx_defaults: key=AppleMeasurementUnits type=string value=Centimeters
 - osx_defaults:
     key: AppleLanguages
@@ -75,7 +83,7 @@ EXAMPLES = '''
 - osx_defaults: domain=com.geekchimp.macable key=ExampleKeyToRemove state=absent
 '''
 
-from datetime import datetime
+import datetime
 
 # exceptions --------------------------------------------------------------- {{{
 class OSXDefaultsException(Exception):
@@ -124,14 +132,16 @@ class OSXDefaults(object):
         if type == "string":
             return str(value)
         elif type in ["bool", "boolean"]:
-            if value.lower() in [True, 1, "true", "1", "yes"]:
+            if isinstance(value, basestring):
+                value = value.lower()
+            if value in [True, 1, "true", "1", "yes"]:
                 return True
-            elif value.lower() in [False, 0, "false", "0", "no"]:
+            elif value in [False, 0, "false", "0", "no"]:
                 return False
             raise OSXDefaultsException("Invalid boolean value: {0}".format(repr(value)))
         elif type == "date":
             try:
-                return datetime.strptime(value.split("+")[0].strip(), "%Y-%m-%d %H:%M:%S")
+                return datetime.datetime.strptime(value.split("+")[0].strip(), "%Y-%m-%d %H:%M:%S")
             except ValueError:
                 raise OSXDefaultsException(
                     "Invalid date value: {0}. Required format yyy-mm-dd hh:mm:ss.".format(repr(value))
@@ -152,6 +162,19 @@ class OSXDefaults(object):
             return value
 
         raise OSXDefaultsException('Type is not supported: {0}'.format(type))
+
+    """ Returns a normalized list of commandline arguments based on the "host" attribute """
+    def _host_args(self):
+        if self.host is None:
+            return []
+        elif self.host == 'currentHost':
+            return ['-currentHost']
+        else:
+            return ['-host', self.host]
+
+    """ Returns a list containing the "defaults" executable and any common base arguments """
+    def _base_command(self):
+        return [self.executable] + self._host_args()
 
     """ Converts array output from defaults to an list """
     @staticmethod
@@ -174,7 +197,7 @@ class OSXDefaults(object):
     """ Reads value of this domain & key from defaults """
     def read(self):
         # First try to find out the type
-        rc, out, err = self.module.run_command([self.executable, "read-type", self.domain, self.key])
+        rc, out, err = self.module.run_command(self._base_command() + ["read-type", self.domain, self.key])
 
         # If RC is 1, the key does not exists
         if rc == 1:
@@ -188,7 +211,7 @@ class OSXDefaults(object):
         type = out.strip().replace('Type is ', '')
 
         # Now get the current value
-        rc, out, err = self.module.run_command([self.executable, "read", self.domain, self.key])
+        rc, out, err = self.module.run_command(self._base_command() + ["read", self.domain, self.key])
 
         # Strip output
         out = out.strip()
@@ -217,7 +240,7 @@ class OSXDefaults(object):
             value = str(self.value)
         elif self.array_add and self.current_value is not None:
             value = list(set(self.value) - set(self.current_value))
-        elif isinstance(self.value, datetime):
+        elif isinstance(self.value, datetime.datetime):
             value = self.value.strftime('%Y-%m-%d %H:%M:%S')
         else:
             value = self.value
@@ -230,14 +253,14 @@ class OSXDefaults(object):
         if not isinstance(value, list):
             value = [value]
 
-        rc, out, err = self.module.run_command([self.executable, 'write', self.domain, self.key, '-' + self.type] + value)
+        rc, out, err = self.module.run_command(self._base_command() + ['write', self.domain, self.key, '-' + self.type] + value)
 
         if rc != 0:
             raise OSXDefaultsException('An error occurred while writing value to defaults: ' + out)
 
     """ Deletes defaults key from domain """
     def delete(self):
-        rc, out, err = self.module.run_command([self.executable, 'delete', self.domain, self.key])
+        rc, out, err = self.module.run_command(self._base_command() + ['delete', self.domain, self.key])
         if rc != 0:
             raise OSXDefaultsException("An error occurred while deleting key from defaults: " + out)
 
@@ -255,6 +278,8 @@ class OSXDefaults(object):
             print "Absent state detected!"
             if self.current_value is None:
                 return False
+            if self.module.check_mode:
+                return True
             self.delete()
             return True
 
@@ -272,6 +297,9 @@ class OSXDefaults(object):
         elif self.current_value == self.value:
             return False
 
+        if self.module.check_mode:
+            return True
+
         # Change/Create/Set given key/value for domain in defaults
         self.write()
         return True
@@ -287,6 +315,10 @@ def main():
         argument_spec=dict(
             domain=dict(
                 default="NSGlobalDomain",
+                required=False,
+            ),
+            host=dict(
+                default=None,
                 required=False,
             ),
             key=dict(
@@ -309,7 +341,7 @@ def main():
             array_add=dict(
                 default=False,
                 required=False,
-                choices=BOOLEANS,
+                type='bool',
             ),
             value=dict(
                 default=None,
@@ -331,6 +363,7 @@ def main():
     )
 
     domain = module.params['domain']
+    host = module.params['host']
     key = module.params['key']
     type = module.params['type']
     array_add = module.params['array_add']
@@ -339,7 +372,7 @@ def main():
     path = module.params['path']
 
     try:
-        defaults = OSXDefaults(module=module, domain=domain, key=key, type=type,
+        defaults = OSXDefaults(module=module, domain=domain, host=host, key=key, type=type,
                                array_add=array_add, value=value, state=state, path=path)
         changed = defaults.run()
         module.exit_json(changed=changed)
